Fussy compiler on another machine...
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@246204 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ funcs/func_strings.c
@@ -638,7 +638,7 @@ static int filter(struct ast_channel *chan, const char *cmd, char *parse, char *
 			ast_debug(4, "c1=%d, c2=%d\n", c1, c2);
 		} else {
 			ac = (unsigned char) c1;
-			ast_debug(4, "c1=%d, consumed=%d, args.allowed=%s\n", c1, consumed, args.allowed - consumed);
+			ast_debug(4, "c1=%d, consumed=%d, args.allowed=%s\n", c1, (int) consumed, args.allowed - consumed);
 			bitfield[ac / 32] |= 1 << (ac % 32);
 		}
 	}
